ocr-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pytesseract==0.3.13
Pillow==11.0.0
python-multipart==0.0.9
tika==3.1.0
